chore: rename include directory meen_i8080_arcade
