+++ include/meen_i8080_arcade/RPIoController.h
@@ -29,7 +29,7 @@ SOFTWARE.
 #include <variant>
 #include <vector>
 
-#include "i8080_arcade/IIoController.h"
+#include "meen_i8080_arcade/IIoController.h"
 #include "meen_hw/MH_Factory.h"
 
 namespace meen_i8080_arcade
